refactor: 전체 패키지명 변경
- Step4 진입에 따른 패키지명 반영 (step3 -> step3_4)
