+++ src/main/java/step3_4/domain/Lotto.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3_4.domain;
 
 public class Lotto {
