+++ src/main/java/step3_4/domain/Rank.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3_4.domain;
 
 import java.util.Arrays;
