+++ src/main/java/step3_4/domain/UniqueNumbers.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3_4.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
