+++ src/test/java/step3_4/domain/LottoStoreTest.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3_4.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
