+++ src/test/java/step3_4/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3_4.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.Arrays;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.utils.NumbersGenerator;
+import step3_4.utils.NumbersGenerator;
 
 public class LottoTest {
