+++ src/test/java/step3_4/domain/UniqueNumbersTest.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3_4.domain;
 
 import java.util.Arrays;
 import java.util.List;
